+++ src/HeatManager/ViewModels/OptimizerGraphs/IOptimizerHeatProductionGraphViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HeatManager.ViewModels.OptimizerGraphs
+{
+    interface IOptimizerHeatProductionGraphViewModel
+    {
+    }
+}
