+++ src/HeatManager/Views/Optimizer/DataOptimizerView.axaml
@@ -3,7 +3,6 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:avalonia="clr-namespace:LiveChartsCore.SkiaSharpView.Avalonia;assembly=LiveChartsCore.SkiaSharpView.Avalonia"
     xmlns:optimizer="clr-namespace:HeatManager.ViewModels.Optimizer"
     mc:Ignorable="d"
     d:DesignWidth="800" d:DesignHeight="450"
@@ -16,87 +15,25 @@
             <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
 
-        <!-- ========================== Header Section ========================== -->
-        <StackPanel Grid.Row="0" Orientation="Horizontal">
-            <TextBlock Text="{Binding DateRangeText}"
-                       FontWeight="SemiBold"
-                       FontSize="14"
-                       Margin="10"
-                       VerticalAlignment="Center"
-                       IsVisible="{Binding IsChartVisible}" />
-
-            <TextBlock Text="|     Navigate to Date:"
-                       FontWeight="SemiBold"
-                       FontSize="14"
-                       Margin="10"
-                       VerticalAlignment="Center"
-                       IsVisible="{Binding IsChartVisible}" />
-
-            <DatePicker Name="DatePickerControl"
-                        SelectedDate="{Binding SelectedDate, Mode=TwoWay}"
-                        HorizontalAlignment="Left"
-                        VerticalAlignment="Top"
-                        Margin="10"
-                        IsVisible="{Binding IsChartVisible}" />
-        </StackPanel>
-
-        <!-- Toggle Button -->
-        <Button Content="Switch View"
+        <!-- Dropdown Button 
+        < Content="Switch View"
                 Margin="10"
                 Click="ToggleView_Click"
                 Grid.Row="0"
-                HorizontalAlignment="Right" />
+                HorizontalAlignment="Right" /> -->
+		<ComboBox SelectedIndex="0" MaxDropDownHeight="100">
+            <ComboBoxItem>Text Item 1</ComboBoxItem>
+            <ComboBoxItem>Text Item 2</ComboBoxItem>
+            <ComboBoxItem>Text Item 3</ComboBoxItem>
+            <ComboBoxItem>Text Item 4</ComboBoxItem>
+            <ComboBoxItem>Text Item 5</ComboBoxItem>
+            <ComboBoxItem>Text Item 6</ComboBoxItem>
+            <ComboBoxItem>Text Item 7</ComboBoxItem>
+            <ComboBoxItem>Text Item 8</ComboBoxItem>
+            <ComboBoxItem>Text Item 9</ComboBoxItem>
+         </ComboBox>
 
-        <!-- ========================== Chart Section ========================== -->
-        <avalonia:CartesianChart Grid.Row="1"
-                                 Series="{Binding Series}"
-                                 XAxes="{Binding XAxes}"
-                                 YAxes="{Binding YAxes}"
-                                 ZoomMode="X"
-                                 TooltipPosition="Top"
-                                 DrawMargin="{Binding Margin}"
-                                 UpdateStartedCommand="{Binding ChartUpdatedCommand}"
-                                 IsVisible="{Binding IsChartVisible}" />
-
-        <!-- ========================== Table Section ========================== -->
-        <Grid Grid.Row="1" IsVisible="{Binding IsTableVisible}">
-            <StackPanel Margin="20">
-                <!-- Total Values Table -->
-                <TextBlock Text="Total Values"
-                           FontWeight="Bold" />
-                
-                <DataGrid Margin="20"
-                          ItemsSource="{Binding TableData}"
-                          IsReadOnly="True"
-                          GridLinesVisibility="All">
-                    <DataGrid.Columns>
-                        <DataGridTextColumn Header="Unit" Binding="{Binding Name}" />
-                        <DataGridTextColumn Header="Heat Production [MW]" Binding="{Binding HeatProduction}" />
-                        <DataGridTextColumn Header="Emissions [kg]" Binding="{Binding Emissions}" />
-                        <DataGridTextColumn Header="Cost [DKK]" Binding="{Binding Cost}" />
-                        <DataGridTextColumn Header="Resource Consumption [MWh]" Binding="{Binding ResourceConsumption}" />
-                        <DataGridTextColumn Header="Utilization [%]" Binding="{Binding Utilization}" />
-                    </DataGrid.Columns>
-                </DataGrid>
-
-                <!-- Maximum Values Table -->
-                <TextBlock Text="Maximum Values"
-                           FontWeight="Bold" />
-
-                <DataGrid Margin="20"
-                          ItemsSource="{Binding TableData}"
-                          IsReadOnly="True"
-                          GridLinesVisibility="All">
-                    <DataGrid.Columns>
-                        <DataGridTextColumn Header="Unit" Binding="{Binding Name}" />
-                        <DataGridTextColumn Header="Heat Production [MW]" Binding="{Binding MaxHeatProduction}" />
-                        <DataGridTextColumn Header="Emissions [kg]" Binding="{Binding MaxEmissions}" />
-                        <DataGridTextColumn Header="Cost [DKK]" Binding="{Binding MaxCost}" />
-                        <DataGridTextColumn Header="Resource Consumption [MWh]" Binding="{Binding MaxResourceConsumption}" />
-                        <DataGridTextColumn Header="Utilization [%]" Binding="{Binding MaxUtilization}" />
-                    </DataGrid.Columns>
-                </DataGrid>
-            </StackPanel>
-        </Grid>
+        <!-- View Selection -->
+			<ContentControl Grid.Row="1" Content = "{Binding CurrentView}" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"/>
     </Grid>
 </UserControl>
